VIDSOL-575 settings module - detekt fixes
