+++ app/src/main/java/com/vonage/android/screen/settings/SettingsScreenRoute.kt
@@ -2,7 +2,6 @@ package com.vonage.android.screen.settings
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
-import androidx.compose.ui.Modifier
 import androidx.hilt.lifecycle.viewmodel.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.vonage.android.BuildConfig
@@ -12,7 +11,6 @@ import com.vonage.android.settings.ui.SettingsScreenViewModelFactory
 
 @Composable
 fun SettingsScreenRoute(
-    modifier: Modifier = Modifier,
     viewModel: SettingsScreenViewModel = hiltViewModel<SettingsScreenViewModel, SettingsScreenViewModelFactory> { factory ->
         factory.create(
             appVersion = BuildConfig.VERSION_NAME,
