+++ app/src/main/java/com/vonage/android/screen/waiting/WaitingRoomScreen.kt
@@ -2,18 +2,13 @@ package com.vonage.android.screen.waiting
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.requiredHeight
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.widthIn
 import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
 import androidx.compose.material3.ModalBottomSheet
-import androidx.compose.material3.Text
 import androidx.compose.material3.rememberModalBottomSheetState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
@@ -24,22 +19,18 @@ import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.PreviewLightDark
 import androidx.compose.ui.tooling.preview.PreviewScreenSizes
 import androidx.compose.ui.unit.dp
-import com.vonage.android.R
 import com.vonage.android.compose.layout.TwoPaneScaffold
 import com.vonage.android.compose.preview.buildPublisher
 import com.vonage.android.compose.theme.VonageVideoTheme
-import com.vonage.android.compose.vivid.icons.VividIcons
-import com.vonage.android.compose.vivid.icons.solid.Gear
-import com.vonage.android.screen.components.TopBanner
 import com.vonage.android.screen.components.audio.AudioDevicesMenu
 import com.vonage.android.screen.waiting.components.DeviceSelectionPanel
 import com.vonage.android.screen.waiting.components.JoinRoomSection
 import com.vonage.android.screen.waiting.components.VideoControlPanel
 import com.vonage.android.screen.waiting.components.VideoPreviewContainer
+import com.vonage.android.screen.waiting.components.WaitingRoomTopBar
 import kotlinx.coroutines.launch
 
 @OptIn(ExperimentalMaterial3Api::class)
@@ -83,28 +74,9 @@ fun WaitingRoomScreen(
     TwoPaneScaffold(
         modifier = modifier.fillMaxSize(),
         topBar = {
-            TopBanner(
-                onBack = actions.onBack,
-                content = {
-                    Text(
-                        text = stringResource(R.string.waiting_room_prepare_to_join),
-                        style = VonageVideoTheme.typography.heading4,
-                        color = VonageVideoTheme.colors.textSecondary,
-                    )
-                },
-                actions = {
-                    IconButton(
-                        modifier = Modifier,
-                        onClick = navigateToSettings,
-                    ) {
-                        Icon(
-                            imageVector = VividIcons.Solid.Gear,
-                            contentDescription = null,
-                            tint = VonageVideoTheme.colors.onSurface,
-                            modifier = modifier.size(VonageVideoTheme.dimens.iconSizeDefault)
-                        )
-                    }
-                }
+            WaitingRoomTopBar(
+                actions = actions,
+                navigateToSettings = navigateToSettings,
             )
         },
         firstPane = {
